refactor(ns04): bound product session closure
