+++ breadboard/product/runtime/__init__.py
@@ -1,4 +1,5 @@
 """Product-owned Session runtime facade."""
 from .artifacts import ArtifactRef, ArtifactStore
-from .events import KernelEvent, Session, SessionView, rebuild
+from .events import KernelEvent, SessionView, rebuild
+from .session import Session
 __all__ = ["ArtifactRef", "ArtifactStore", "KernelEvent", "Session", "SessionView", "rebuild"]
