+++ tests/test_session_runner_pending_permissions.py
@@ -1,148 +1,72 @@
 from __future__ import annotations
-import asyncio, json, multiprocessing, os, stat, threading, pytest; from pathlib import Path; from breadboard.product.harness.lock import EffectiveHarnessLock; from breadboard.product.runtime.events import Session as ProductSession
-from agentic_coder_prototype.api.cli_bridge.events import EventType; from agentic_coder_prototype.permissions import load_permission_rules, upsert_permission_rule; from agentic_coder_prototype.permissions import rules_store; from agentic_coder_prototype.permissions.broker import PermissionBroker; from agentic_coder_prototype.permissions.rules_store import RULES_REL_PATH, _locked_rules
-from agentic_coder_prototype.api.cli_bridge.models import SessionCreateRequest, SessionStatus; from agentic_coder_prototype.api.cli_bridge.registry import SessionRecord, SessionRegistry; from agentic_coder_prototype.api.cli_bridge.session_runner import SessionRunner, _canonical_permission_resolution
-from agentic_coder_prototype.api.cli_bridge.service import SessionService
-def _runner(session_id: str = "session") -> SessionRunner: return SessionRunner(session=SessionRecord(session_id=session_id, status=SessionStatus.RUNNING), registry=SessionRegistry(), request=SessionCreateRequest(config_path="dummy.yml", task="task", stream=False))
-def _product_runner(session_id: str) -> tuple[SessionRunner, ProductSession]: runner = _runner(session_id); session = ProductSession.start(EffectiveHarnessLock._from_record({"graph_hash": "sha256:" + "a" * 64}), "task", session_id=session_id); runner.session.product_session = session; return runner, session
-async def _initialized() -> None: pass
-def _upsert_process(workspace: str, pattern: str) -> None: assert upsert_permission_rule(Path(workspace), category="shell", pattern=pattern, decision="allow")
-def _hold_rule_lock(workspace: str, entered, release) -> None: lock = _locked_rules(Path(workspace) / RULES_REL_PATH); lock.__enter__(); entered.set(); release.wait(10); lock.__exit__(None, None, None)  # type: ignore[no-untyped-def]
-def test_pending_permissions_rehydrate_and_remain_scoped() -> None:
-    runner = _runner(); runner._rehydrate_pending_permissions("permission_request", {"request_id": "session", "items": []}); task = {"kind": "permission_request", "sessionId": "task-1", "subagent_type": "general", "payload": {"request_id": "task", "items": []}}; runner._rehydrate_pending_permissions("task_event", task); assert [(item["source"], item.get("task_session_id"), item.get("subagent_type"), item["request_id"]) for item in runner.session.metadata["pending_permissions"]] == [("session", None, None, "session"), ("task", "task-1", "general", "task")]
-    task.update(kind="permission_response", payload={"request_id": "task", "responses": {"default": "once"}}); runner._rehydrate_pending_permissions("task_event", task); assert len(runner.session.metadata["pending_permissions"]) == 2; runner._rehydrate_pending_permissions("permission_response", {"request_id": "session", "responses": {"default": "once"}}); runner._rehydrate_pending_permissions("task_event", task); assert "pending_permissions" not in runner.session.metadata
-    for source, task_id in (("task", "a"), ("session", None), ("task", "b")): runner._update_pending_permissions("permission_request", {"request_id": "shared", "items": []}, source=source, task_session_id=task_id)
-    runner._update_pending_permissions("permission_response", {"request_id": "shared", "response": "once"}, source="task", task_session_id="a"); assert [runner._pending_permission_key(item) for item in runner.session.metadata["pending_permissions"]] == [("session", "", "shared"), ("task", "b", "shared")]
-@pytest.mark.parametrize(("expected", "aliases"), [("once", "allow approve approved ok okay yes y allow-once allow_once once"), ("always", "allow-always allow_always always"), ("reject", "deny denied no n deny-once deny_once deny-always deny_always deny-stop deny_stop reject")])
-def test_permission_aliases_match_broker(expected: str, aliases: str) -> None: assert {_canonical_permission_resolution(alias, None) for alias in aliases.split()} == {expected}
-def test_unknown_permission_decision_matches_broker_reject() -> None: assert _canonical_permission_resolution("garbage", None) == PermissionBroker._coerce_response("garbage") == "reject"
-def test_attachment_reads_enter_permission_broker_without_changing_workspace_reads() -> None: uri = "attachment://sha256:" + "a" * 64; attachment = type("Call", (), {"function": "read_file", "arguments": {"path": uri}})(); workspace = type("Call", (), {"function": "read_file", "arguments": {"path": "README.md"}})(); assert PermissionBroker({"read": {"default": "ask"}}).decide(attachment) == "ask" and PermissionBroker({"read": {"default": "deny"}}).decide(attachment) == "deny" and PermissionBroker().decide(workspace) is None
-def test_explicit_item_allows_ignore_reject_fallback_until_an_item_is_missing() -> None: responses = {"items": {"a": "always", "b": "always"}, "fallback": "reject"}; assert _canonical_permission_resolution(None, responses, ["a", "b"]) == "always" and _canonical_permission_resolution(None, responses, ["a", "b", "c"]) == "reject" and _canonical_permission_resolution(None, {"a": "always"}, ["a", "b"]) == "reject" and _canonical_permission_resolution(None, {"default": "always", "items": {"a": "reject"}}, ["a"]) == "always"
-def test_runtime_permission_replay_uses_configured_default_for_sparse_items() -> None: runner, session = _product_runner("batch"); config = {"options": {"default_response": "always"}}; runner._agent = type("Agent", (), {"config": {"permissions": config}})(); runner._rehydrate_pending_permissions("permission_request", {"request_id": "batch", "items": [{"item_id": "a"}, {"item_id": "b"}]}); response = {"responses": {"a": "always"}}; runner._rehydrate_pending_permissions("permission_response", {"request_id": "batch", **response}); assert set(PermissionBroker(config)._parse_permission_response_item(response, request_id="batch", item_ids=["a", "b"]).values()) == {"always"} and session.events[-1].payload["decision"] == "always"
-@pytest.mark.asyncio
-@pytest.mark.parametrize("echo_order", [("permission_response", "task_event"), ("task_event", "permission_response")])
-async def test_command_then_scoped_broker_echo_consumes_fifo_entry_once(echo_order: tuple[str, str]) -> None: runner, session = _product_runner("echo"); task = {"kind": "permission_request", "sessionId": "task-a", "payload": {"request_id": "shared", "items": []}}; runner._rehydrate_pending_permissions("task_event", task); runner._rehydrate_pending_permissions("permission_request", {"request_id": "shared", "items": []}); runner._permission_queue = asyncio.Queue(); await runner.handle_command("respond_permission", {"request_id": "shared", "response": "allow_once"}); echoes = {"permission_response": ("permission_response", {"request_id": "shared", "response": "once"}), "task_event": ("task_event", {**task, "kind": "permission_response", "payload": {"request_id": "shared", "response": "once"}})}; [runner._rehydrate_pending_permissions(*echoes[kind]) for kind in echo_order]; assert "pending_permissions" not in runner.session.metadata and not runner._consumed_permission_responses; assert [event.kind for event in session.events[1:]] == ["approval.requested", "approval.resolved", "approval.requested", "approval.resolved"]
-@pytest.mark.asyncio
-async def test_failing_approval_sink_suppresses_pending_and_bridge_event() -> None:
-    runner, session = _product_runner("sink-failure"); session._sink = type("Failing", (), {"append": lambda *_: (_ for _ in ()).throw(OSError("sink unavailable"))})()
-    with pytest.raises(OSError, match="sink unavailable"): await runner._emit_debug_permission_request({"request_id": "undurable"})
-    assert "pending_permissions" not in runner.session.metadata; assert runner.session.event_queue.empty()
-@pytest.mark.asyncio
-@pytest.mark.parametrize(("payload", "expected"), [({"response": "allow-once"}, {"a": "once", "b": "once"}), ({"responses": {"items": {"a": "allow_once", "b": "allow-always"}}}, {"a": "once", "b": "always"})])
-async def test_real_broker_receives_the_persisted_canonical_decision(payload: dict[str, object], expected: dict[str, str]) -> None: runner, session = _product_runner("broker-" + expected["b"]); runner._rehydrate_pending_permissions("permission_request", {"request_id": "permission-1", "items": []}); runner._permission_queue = asyncio.Queue(); delivered = (await runner.handle_command("respond_permission", {"request_id": "permission-1", **payload}))["delivered"]; actual = PermissionBroker()._parse_permission_response_item(delivered, request_id="permission-1", item_ids=["a", "b"]); assert actual == expected; assert session.events[-1].payload["decision"] == _canonical_permission_resolution(None, {"items": actual})
-@pytest.mark.asyncio
-async def test_durable_rule_remains_authoritative_when_metadata_projection_fails(monkeypatch, tmp_path) -> None:
-    runner, session = _product_runner("always"); runner._workspace_path = tmp_path; runner._rehydrate_pending_permissions("permission_request", {"request_id": "permission-1", "category": "shell"}); runner._permission_queue = asyncio.Queue(); fsyncs, real_fsync = [], os.fsync; monkeypatch.setattr(os, "fsync", lambda fd: fsyncs.append(fd) or real_fsync(fd)); assert upsert_permission_rule(tmp_path, category="shell", pattern="safe.sh", decision="allow")
-    async def fail(*_a, **_kw): raise OSError("metadata unavailable")  # type: ignore[no-untyped-def]
-    monkeypatch.setattr(runner.registry, "update_metadata", fail)
-    result = await runner.handle_command("permission_decision", {"request_id": "permission-1", "decision": "always", "rule": "*.sh"})
-    assert fsyncs and result["decision"] == "always" and [(rule.pattern, rule.decision) for rule in load_permission_rules(tmp_path)] == [("safe.sh", "allow"), ("*.sh", "allow")]
-    assert runner.session.metadata["permission_rules"][0]["rule"] == "*.sh" and session.read_model.status == "running" and runner._permission_queue.qsize() == 1
-    runner._rehydrate_pending_permissions("permission_request", {"request_id": "permission-2", "category": "shell"}); runner._permission_queue = None; failure = (await asyncio.gather(runner.handle_command("permission_decision", {"request_id": "permission-2", "decision": "always", "rule": "missing.sh"}), return_exceptions=True))[0]
-    assert isinstance(failure, RuntimeError) and session.read_model.status == "failed"
-    workers = [threading.Thread(target=upsert_permission_rule, kwargs={"workspace_dir": tmp_path, "category": "shell", "pattern": f"parallel-{index}.sh", "decision": "allow"}) for index in range(4)]; [worker.start() for worker in workers]; [worker.join() for worker in workers]; assert {rule.pattern for rule in load_permission_rules(tmp_path)}.issuperset({f"parallel-{index}.sh" for index in range(4)})
-    processes = [multiprocessing.get_context("spawn").Process(target=_upsert_process, args=(str(tmp_path), f"process-{index}.sh")) for index in range(4)]; [process.start() for process in processes]; [process.join(10) for process in processes]; assert all(process.exitcode == 0 for process in processes) and {rule.pattern for rule in load_permission_rules(tmp_path)}.issuperset({f"process-{index}.sh" for index in range(4)})
-    context = multiprocessing.get_context("spawn"); entered, release = context.Event(), context.Event(); holder = context.Process(target=_hold_rule_lock, args=(str(tmp_path), entered, release)); holder.start(); assert entered.wait(5); writer = context.Process(target=_upsert_process, args=(str(tmp_path), "blocked.sh")); writer.start(); writer.join(.2); assert writer.is_alive(); release.set(); holder.join(10); writer.join(10); assert holder.exitcode == writer.exitcode == 0
-@pytest.mark.skipif(os.name == "nt", reason="symlink creation requires privileges on Windows")
-def test_permission_rules_reject_workspace_metadata_symlink(monkeypatch, tmp_path) -> None:
-    outside = tmp_path / "outside"; outside.mkdir(); metadata = tmp_path / ".breadboard"; metadata.symlink_to(outside, target_is_directory=True)
-    assert load_permission_rules(tmp_path) == []; pytest.raises(OSError, upsert_permission_rule, tmp_path, category="shell", pattern="escape.sh", decision="allow"); assert not list(outside.iterdir()); metadata.unlink(); metadata.mkdir(); lock_target = outside / "lock"; lock_path = metadata / "permission_rules.json.lock"; lock_path.symlink_to(lock_target); pytest.raises(OSError, upsert_permission_rule, tmp_path, category="shell", pattern="lock.sh", decision="allow"); assert not lock_target.exists(); lock_path.unlink()
-    real_write, old = rules_store._write_at, tmp_path / ".breadboard-old"
-    def swap_write(*args, **kwargs): metadata.rename(old); metadata.symlink_to(outside, target_is_directory=True); return real_write(*args, **kwargs)  # type: ignore[no-untyped-def]
-    monkeypatch.setattr(rules_store, "_write_at", swap_write); assert upsert_permission_rule(tmp_path, category="shell", pattern="anchored.sh", decision="allow")
-    assert not list(outside.iterdir()) and b"anchored.sh" in (old / "permission_rules.json").read_bytes()
-@pytest.mark.asyncio
-async def test_directory_fsync_failure_never_releases_always_decision(monkeypatch, tmp_path) -> None:
-    runner, session = _product_runner("fsync"); runner._workspace_path = tmp_path; runner._rehydrate_pending_permissions("permission_request", {"request_id": "permission-1", "category": "shell"}); runner._permission_queue = asyncio.Queue(); real_fsync = os.fsync; monkeypatch.setattr(os, "fsync", lambda fd: (_ for _ in ()).throw(OSError("directory fsync failed")) if stat.S_ISDIR(os.fstat(fd).st_mode) else real_fsync(fd))
-    with pytest.raises(RuntimeError, match="failed to commit permission decision"): await runner.handle_command("permission_decision", {"request_id": "permission-1", "decision": "always", "rule": "*.sh"})
-    assert runner._permission_queue.empty() and session.read_model.status == "failed"
-@pytest.mark.asyncio
-async def test_only_active_permission_can_commit_and_concurrent_duplicates_deliver_once() -> None:
-    runner, _ = _product_runner("active"); runner._rehydrate_pending_permissions("permission_request", {"request_id": "first"}); runner._rehydrate_pending_permissions("permission_request", {"request_id": "second"}); runner._permission_queue = asyncio.Queue()
-    with pytest.raises(ValueError, match="not active"): await runner.handle_command("permission_decision", {"request_id": "second", "decision": "once"})
-    results = await asyncio.gather(*(runner.handle_command("permission_decision", {"request_id": "first", "decision": "once"}) for _ in range(2)), return_exceptions=True)
-    assert sum(isinstance(item, dict) for item in results) == 1 and sum(isinstance(item, ValueError) for item in results) == 1 and runner._permission_queue.qsize() == 1
-def test_runtime_response_is_buffered_and_published_in_fifo_order() -> None:
-    runner, session = _product_runner("provider")
-    def run_task(*_a, **kw): [kw["event_emitter"]("task_event", {"kind": kind, "sessionId": task_id, "payload": {"request_id": request_id, **({"response": "once"} if kind == "permission_response" else {})}}) for kind, task_id, request_id in (("permission_request", "task-a", "first"), ("permission_request", "task-b", "second"), ("permission_response", "task-b", "second"), ("permission_response", "task-a", "first"))]; return {"completion_summary": {"completed": True}}  # type: ignore[no-untyped-def]
-    runner._agent = type("Agent", (), {"_local_mode": True, "config": {}, "run_task": run_task})(); runner._execute_task("task"); responses = [event.payload["payload"]["request_id"] for event in runner.session.event_queue._queue if event and event.type is EventType.TASK_EVENT and event.payload.get("kind") == "permission_response"]; assert responses == ["first", "second"] and "pending_permissions" not in runner.session.metadata; assert [(event.kind, event.payload.get("request_id")) for event in session.events[1:]] == [("approval.requested", "first"), ("approval.resolved", "first"), ("approval.requested", "second"), ("approval.resolved", "second")]
-def test_legacy_runtime_response_replays_deferred_fifo_entry() -> None:
-    runner = _runner("legacy-provider"); runner._agent = type("Agent", (), {"_local_mode": True, "config": {}, "run_task": lambda *_a, **kw: ([kw["event_emitter"]("task_event", {"kind": kind, "sessionId": task_id, "payload": {"request_id": request_id, **({"response": "once"} if kind == "permission_response" else {})}}) for kind, task_id, request_id in (("permission_request", "task-a", "first"), ("permission_request", "task-b", "second"), ("permission_response", "task-b", "second"), ("permission_response", "task-a", "first"))], {"completion_summary": {"completed": True}})[1]})(); runner._execute_task("task"); responses = [event.payload["payload"]["request_id"] for event in runner.session.event_queue._queue if event and event.type is EventType.TASK_EVENT and event.payload.get("kind") == "permission_response"]; assert responses == ["first", "second"] and "pending_permissions" not in runner.session.metadata
-def test_runtime_response_resolves_fifo_head_before_projection() -> None:
-    runner, session = _product_runner("provider-failure"); runner._rehydrate_pending_permissions("permission_request", {"request_id": "first"}); runner._rehydrate_pending_permissions("permission_request", {"request_id": "second"}); sink = session._sink; session._sink = type("Failing", (), {"append": lambda *_: (_ for _ in ()).throw(OSError("sink unavailable"))})()
-    with pytest.raises(OSError, match="sink unavailable"): runner._rehydrate_pending_permissions("permission_response", {"request_id": "first", "response": "once"})
-    assert [item["request_id"] for item in runner.session.metadata["pending_permissions"]] == ["first", "second"]; session._sink = type("FailNext", (), {"append": lambda _, event: (_ for _ in ()).throw(OSError("activation unavailable")) if event.kind == "approval.requested" else None})()
-    with pytest.raises(OSError, match="activation unavailable"): runner._rehydrate_pending_permissions("permission_response", {"request_id": "first", "response": "once"})
-    assert [item["request_id"] for item in runner.session.metadata["pending_permissions"]] == ["second"] and session.read_model.status == "running"; session._sink = sink; runner._rehydrate_pending_permissions("permission_request", {"request_id": "second"}); assert [(event.kind, event.payload.get("request_id")) for event in session.events[-2:]] == [("approval.resolved", "first"), ("approval.requested", "second")]
-@pytest.mark.asyncio
-async def test_concurrent_set_mode_failure_wins_oneshot_completion(monkeypatch) -> None:
-    runner, session = _product_runner("mode"); calls, committed, started, release = [], [True], threading.Event(), threading.Event(); runner.session.metadata["cli_session_kind"] = "oneshot"; await runner.registry.create(runner.session); runner._agent = type("Agent", (), {"_local_mode": True, "config": {"mode": "old"}, "apply_runtime_overrides": lambda self, value: (calls.append(("apply", value["mode"])), self.config.update(value) if committed[0] else None, committed[0])[-1], "run_task": lambda *_a, **_kw: (started.set(), release.wait(timeout=1), {"completion_summary": {"completed": True}})[-1]})(); await runner.handle_command("set_mode", {"mode": "review"}, durable_reconfigure=lambda cfg: calls.append(("append", cfg["mode"]))); committed[0] = False; monkeypatch.setattr(runner, "prepare_runtime_config", lambda: {}); monkeypatch.setattr(runner, "_ensure_agent_initialized", _initialized); running = asyncio.create_task(runner._run()); assert await asyncio.to_thread(started.wait, 1)
-    with pytest.raises(RuntimeError, match="failed to apply mode"): await runner.handle_command("set_mode", {"mode": "broken"}, durable_reconfigure=lambda cfg: calls.append(("append", cfg["mode"])))
-    release.set(); await running; visible = {event.type for event in runner.session.event_queue._queue if event}; assert calls[-2:] == [("apply", "broken"), ("apply", "review")] and runner.session.metadata["mode"] == "review"; assert (runner.session.status, session.read_model.status, runner.session.completion_summary) == (SessionStatus.FAILED, "failed", None) and not visible.intersection({EventType.COMPLETION, EventType.RUN_FINISHED})
-@pytest.mark.asyncio
-async def test_stop_wakes_oneshot_and_wins_terminal_status(monkeypatch) -> None:
-    runner, session = _product_runner("stopped"); runner.session.metadata["cli_session_kind"] = "oneshot"; await runner.registry.create(runner.session); monkeypatch.setattr(runner, "prepare_runtime_config", lambda: {}); monkeypatch.setattr(runner, "_ensure_agent_initialized", _initialized); monkeypatch.setattr(runner, "_execute_task", lambda _task: (runner._request_stop("race"), {"completion_summary": {"completed": True}})[1]); await runner._run()
-    assert (runner.session.status, session.read_model.status, runner._stop_event.is_set()) == (SessionStatus.STOPPED, "canceled", True); assert None in runner._input_queue._queue
-def test_stop_signals_runtime_when_cancel_evidence_fails() -> None:
-    runner, session = _product_runner("stop-sink"); runner._control_queue = asyncio.Queue(); session._sink = type("Failing", (), {"append": lambda *_: (_ for _ in ()).throw(OSError("sink unavailable"))})()
-    with pytest.raises(OSError, match="sink unavailable"): runner._request_stop("operator")
-    assert runner._stop_event.is_set() and runner._resume_event.is_set() and runner._input_queue.get_nowait() is None and runner._control_queue.get_nowait() == {"kind": "stop"} and session.read_model.status == "running"
-@pytest.mark.asyncio
-async def test_stop_during_agent_initialization_never_executes_task(monkeypatch) -> None:
-    runner, session = _product_runner("init-stop"); runner.session.metadata["cli_session_kind"] = "oneshot"; await runner.registry.create(runner.session); calls = []; monkeypatch.setattr(runner, "prepare_runtime_config", lambda: {}); monkeypatch.setattr(runner, "_ensure_agent_initialized", lambda: asyncio.sleep(0, result=runner._request_stop("init"))); monkeypatch.setattr(runner, "_execute_task", lambda task: calls.append(task)); await runner._run(); assert calls == [] and session.read_model.status == "canceled"
-@pytest.mark.asyncio
-async def test_pause_blocks_dequeued_work_until_resume_and_stop_signals_active_agent(monkeypatch) -> None:
-    runner, session = _product_runner("control"); runner.session.metadata["cli_session_kind"] = "oneshot"; await runner.registry.create(runner.session); started, paused, stopped = threading.Event(), threading.Event(), threading.Event()
-    def run_task(*_a, **kw): started.set(); assert kw["control_queue"].get(timeout=1) == {"kind": "pause"}; paused.set(); assert kw["control_queue"].get(timeout=1) == {"kind": "resume"}; stopped.set() if kw["control_queue"].get(timeout=1) == {"kind": "stop"} else None; return {"completion_summary": {"completed": True}}  # type: ignore[no-untyped-def]
-    runner._agent = type("Agent", (), {"_local_mode": True, "config": {}, "run_task": run_task, "apply_runtime_overrides": lambda *_: True})(); monkeypatch.setattr(runner, "prepare_runtime_config", lambda: {}); monkeypatch.setattr(runner, "_ensure_agent_initialized", _initialized)
-    await runner.handle_command("pause", {}); running = asyncio.create_task(runner._run()); await asyncio.sleep(0.05); assert not started.is_set() and session.read_model.status == "paused"
-    await runner.handle_command("resume", {}); assert await asyncio.to_thread(started.wait, 1); await runner.handle_command("pause", {}); assert await asyncio.to_thread(paused.wait, 1) and session.read_model.status == "paused"; await runner.handle_command("resume", {}); runner._request_stop("operator"); await running; assert stopped.is_set() and session.read_model.status == "canceled"
-@pytest.mark.asyncio
-async def test_legacy_oneshot_runner_without_product_session_completes(monkeypatch) -> None:
-    runner = _runner("legacy"); runner.session.metadata["cli_session_kind"] = "oneshot"; await runner.registry.create(runner.session); monkeypatch.setattr(runner, "prepare_runtime_config", lambda: {}); monkeypatch.setattr(runner, "_ensure_agent_initialized", _initialized); monkeypatch.setattr(runner, "_execute_task", lambda _task: {"completion_summary": {"completed": True}}); await runner._run()
-    record = await runner.registry.get(runner.session.session_id); assert record is not None and (record.status, record.completion_summary) == (SessionStatus.COMPLETED, {"completed": True}) and None in runner.session.event_queue._queue; assert not any(event and event.type is EventType.ERROR for event in runner.session.event_queue._queue)
-@pytest.mark.asyncio
-async def test_oneshot_success_is_hidden_until_durable_complete(monkeypatch) -> None:
-    runner, session = _product_runner("completion"); runner.session.metadata["cli_session_kind"] = "oneshot"; await runner.registry.create(runner.session); monkeypatch.setattr(runner, "prepare_runtime_config", lambda: {}); runner._agent = type("Agent", (), {"_local_mode": True, "config": {}, "run_task": lambda *_a, **kw: (kw["event_emitter"]("completion", {"summary": {}}), kw["event_emitter"]("run_finished", {"completed": True}), {"completion_summary": {"completed": True}})[-1]})(); monkeypatch.setattr(runner, "_ensure_agent_initialized", _initialized); session._sink = type("Failing", (), {"append": lambda *_: (_ for _ in ()).throw(OSError("sink unavailable"))})()
-    with pytest.raises(OSError, match="sink unavailable"): await runner._run()
-    visible = {event.type for event in runner.session.event_queue._queue if event is not None}; assert not visible.intersection({EventType.COMPLETION, EventType.RUN_FINISHED}) and runner.session.completion_summary is None
-@pytest.mark.asyncio
-async def test_post_complete_metadata_failure_keeps_terminal_authority(monkeypatch) -> None:
-    runner, session = _product_runner("post-complete"); observed = []; runner.session.metadata["cli_session_kind"] = "oneshot"; await runner.registry.create(runner.session); runner._agent = type("Agent", (), {"_local_mode": True, "config": {}, "run_task": lambda *_a, **kw: (kw["event_emitter"]("completion", {"summary": {"provider": True}}), kw["event_emitter"]("run_finished", {"completed": True}), {"completion_summary": {"completed": True}})[-1]})(); monkeypatch.setattr(runner, "prepare_runtime_config", lambda: {}); monkeypatch.setattr(runner, "_ensure_agent_initialized", _initialized)
-    async def fail(*_a, **_kw): observed.append(session.read_model.status); raise OSError("metadata unavailable")  # type: ignore[no-untyped-def]
-    monkeypatch.setattr(runner.registry, "update_metadata", fail); await runner._run(); record = await runner.registry.get(runner.session.session_id); visible = {event.type for event in runner.session.event_queue._queue if event}; assert observed == ["completed"] and record is not None and (record.status, session.read_model.status) == (SessionStatus.COMPLETED, "completed") and not visible.intersection({EventType.ERROR, EventType.COMPLETION, EventType.RUN_FINISHED}) and None in runner.session.event_queue._queue
-@pytest.mark.asyncio
-@pytest.mark.parametrize("one_shot", [False, True])
-async def test_provider_terminal_events_are_payload_faithful_and_exactly_once(monkeypatch, one_shot: bool) -> None:
-    runner, _ = _product_runner("terminal"); completion, finished = {"summary": {"provider": True}, "opaque": "completion"}, {"completed": True, "opaque": "finished"}; barrier = threading.Barrier(2)
-    def run_task(*_a, **kw): emit = lambda: (barrier.wait(timeout=1), kw["event_emitter"]("completion", completion, turn=7), barrier.wait(timeout=1), kw["event_emitter"]("run_finished", finished, turn=8)); threads = [threading.Thread(target=emit, daemon=True) for _ in range(2)]; [thread.start() for thread in threads]; [thread.join(timeout=2) for thread in threads]; assert all(not thread.is_alive() for thread in threads); return {"completion_summary": {"completed": True}}  # type: ignore[no-untyped-def]
-    runner._agent = type("Agent", (), {"_local_mode": True, "config": {}, "run_task": run_task})(); runner.session.metadata.update({"cli_session_kind": "oneshot"} if one_shot else {}); monkeypatch.setattr(runner, "prepare_runtime_config", lambda: {}); monkeypatch.setattr(runner, "_ensure_agent_initialized", _initialized); _ = await runner.registry.create(runner.session) if one_shot else runner._execute_task("task"); await runner._run() if one_shot else None; assert [(event.type, event.payload, event.turn, event.classification, event.family, event.actor, event.visibility) for event in runner.session.event_queue._queue if event and event.type in {EventType.COMPLETION, EventType.RUN_FINISHED}] == [(EventType.COMPLETION, completion, 7, "bridge_host", "host.completion", "service", "host"), (EventType.RUN_FINISHED, finished, 8, "bridge_host", "host.run_finished", "service", "host")]
-class _GateLock:
-    def __init__(self) -> None: self.lock, self.gate, self.attempts = threading.RLock(), threading.Barrier(2), 0
-    def __enter__(self) -> "_GateLock": self.attempts += 1; self.gate.wait(timeout=1); self.lock.acquire(); return self
-    def __exit__(self, *_args: object) -> None: self.lock.release()
-def test_runner_serializes_concurrent_product_transitions() -> None:
-    runner, session = _product_runner("concurrent"); gate, errors = _GateLock(), []; runner._product_session_lock = gate
-    def submit(content: str) -> None:
-        try: asyncio.run(runner.enqueue_input(content))
-        except BaseException as error: errors.append(error)
-    threads = [threading.Thread(target=submit, args=(content,), daemon=True) for content in ("first", "second")]; [thread.start() for thread in threads]; [thread.join(timeout=2) for thread in threads]
-    assert gate.attempts == 2 and not errors and all(not thread.is_alive() for thread in threads); assert ([event.sequence for event in session.events], runner._input_queue.qsize()) == ([1, 2, 3], 2)
-@pytest.mark.asyncio
-async def test_mutating_loader_cannot_split_frozen_start_artifacts(monkeypatch, tmp_path: Path) -> None:
-    first = {"tools": {"registry": {"include": ["first_tool"]}}, "policies": {"tools": {"allow": ["first_tool"], "deny": ["blocked_tool"]}, "models": {"deny": ["blocked-model"]}, "approvals": {"mode": "always_required"}}, "provider_auth_runtime": {"openai": {"api_key": "runtime-secret"}}, "provider_auth_runtime.openai.api_key": "flat-secret", "wrapper": {"provider_auth_runtime": {"openai": {"api_key": "nested-secret"}}}}
-    source, calls = json.loads(json.dumps(first)), []
-    def load(path): calls.append(path); return source  # type: ignore[no-untyped-def]
-    def enable_primitives() -> bool: source["tools"]["registry"]["include"] = ["second_tool"]; source["policies"]["tools"] = {"allow": ["second_tool"], "deny": []}; source["policies"]["models"] = {"deny": []}; return True
-    async def start(_runner) -> None: return None  # type: ignore[no-untyped-def]
-    monkeypatch.setattr("agentic_coder_prototype.api.cli_bridge.session_runner.load_agent_config", load); monkeypatch.setattr("agentic_coder_prototype.api.cli_bridge.runtime_emission.load_agent_config", lambda _: pytest.fail("emitter reloaded mutable source")); monkeypatch.setattr("agentic_coder_prototype.api.cli_bridge.service.primitive_emission_enabled", enable_primitives)
-    monkeypatch.setattr("agentic_coder_prototype.api.cli_bridge.service.uuid.uuid4", lambda: "frozen-session"); monkeypatch.setattr("agentic_coder_prototype.api.cli_bridge.session_runner.SessionRunner.start", start); monkeypatch.setenv("BREADBOARD_RUNTIME_RECORD_ROOT", str(tmp_path / "records")); monkeypatch.setenv("BREADBOARD_SESSION_EVENT_ROOT", str(tmp_path / "events"))
-    service = SessionService(); response = await service.create_session(SessionCreateRequest(config_path="mutating.json", task="test")); record = await service.ensure_session(response.session_id); root = tmp_path / "records" / response.session_id; names = ("effective_config_graph", "capability_registry", "effective_tool_surface", "effective_operation_policy"); payloads = {name: json.loads((root / f"{name}.json").read_text()) for name in names}
-    graph, policy = payloads["effective_config_graph"], payloads["effective_operation_policy"]; tool_ids = {item["capability_id"] for item in payloads["capability_registry"]["capabilities"] if item["capability_type"] == "tool"}; assert calls == ["mutating.json"] and graph["graph_hash"] == record.product_session.events[0].payload["effective_lock_hash"]
-    assert tool_ids == {"tool.first_tool"} == set(payloads["effective_tool_surface"]["tool_ids"]) and [(rule["decision"], rule["match"]["pattern"]) for rule in policy["tool_policy"]["rules"]] == [("allow", "first_tool"), ("deny", "blocked_tool")]
-    serialized = json.dumps(payloads) + "".join(path.read_text() for path in tmp_path.rglob("*") if path.is_file()); assert policy["approvals"]["mode"] == "always_required" and all(secret not in serialized for secret in ("provider_auth_runtime", "runtime-secret", "flat-secret", "nested-secret", "second_tool"))
-    captured = {}; Agent = type("Agent", (), {"workspace_dir": str(tmp_path / "workspace"), "initialize": lambda self: None}); factory = lambda path, _workspace, _overrides: (captured.update(config=json.loads(Path(path).read_text()), path=path), Agent())[1]  # type: ignore[assignment]
-    record.runner.agent_factory = factory; record.runner.get_skill_catalog(); await record.runner._ensure_agent_initialized()
-    persisted = "".join(path.read_text() for root in (tmp_path / "records", tmp_path / "events") for path in root.rglob("*") if path.is_file()); assert calls == ["mutating.json"] and captured["config"] == record.runner.current_runtime_config() and captured["config"]["provider_auth_runtime"]["openai"]["api_key"] == "runtime-secret" and all(secret not in persisted for secret in ("provider_auth_runtime", "runtime-secret", "flat-secret", "nested-secret")) and not Path(captured["path"]).exists()
-    with pytest.raises(ValueError, match="denied by policy"): await record.runner.handle_command("set_model", {"model": "blocked-model"})
-    await service.stop_session(response.session_id); await record.event_queue.put(None); await record.dispatcher_task
+
+from agentic_coder_prototype.api.cli_bridge.models import SessionCreateRequest, SessionStatus
+from agentic_coder_prototype.api.cli_bridge.registry import SessionRecord, SessionRegistry
+from agentic_coder_prototype.api.cli_bridge.session_runner import SessionRunner
+
+
+def test_session_runner_tracks_pending_permissions_for_rehydration() -> None:
+    registry = SessionRegistry()
+    session = SessionRecord(session_id="sess_1", status=SessionStatus.RUNNING)
+    request = SessionCreateRequest(config_path="dummy.yml", task="hi", stream=False)
+    runner = SessionRunner(session=session, registry=registry, request=request)
+
+    runner._rehydrate_pending_permissions(
+        "permission_request",
+        {
+            "event_version": 1,
+            "request_id": "permission_1",
+            "items": [
+                {"item_id": "permission_1_item_1", "category": "shell", "pattern": "echo hi", "metadata": {}},
+            ],
+        },
+    )
+
+    pending = session.metadata.get("pending_permissions")
+    assert isinstance(pending, list) and pending
+    assert pending[0].get("source") == "session"
+    assert pending[0].get("request_id") == "permission_1"
+
+    runner._rehydrate_pending_permissions("permission_response", {"request_id": "permission_1", "responses": {"default": "once"}})
+    assert "pending_permissions" not in session.metadata
+
+
+def test_session_runner_tracks_task_permission_requests_separately() -> None:
+    registry = SessionRegistry()
+    session = SessionRecord(session_id="sess_2", status=SessionStatus.RUNNING)
+    request = SessionCreateRequest(config_path="dummy.yml", task="hi", stream=False)
+    runner = SessionRunner(session=session, registry=registry, request=request)
+
+    runner._rehydrate_pending_permissions(
+        "task_event",
+        {
+            "kind": "permission_request",
+            "sessionId": "task_abc",
+            "subagent_type": "general",
+            "payload": {
+                "event_version": 1,
+                "request_id": "permission_2",
+                "items": [
+                    {"item_id": "permission_2_item_1", "category": "shell", "pattern": "echo hi", "metadata": {}},
+                ],
+            },
+        },
+    )
+
+    pending = session.metadata.get("pending_permissions")
+    assert isinstance(pending, list) and pending
+    assert pending[0].get("source") == "task"
+    assert pending[0].get("task_session_id") == "task_abc"
+    assert pending[0].get("request_id") == "permission_2"
+
+    runner._rehydrate_pending_permissions(
+        "task_event",
+        {
+            "kind": "permission_response",
+            "sessionId": "task_abc",
+            "subagent_type": "general",
+            "payload": {"request_id": "permission_2", "responses": {"default": "once"}},
+        },
+    )
+    assert "pending_permissions" not in session.metadata
+
